Optimise psFFT squaring and hence sisFFT.
When we're squaring we can optimise a little over the general case:

- only need to compute splits and ffts once
- only need to do one triangle of the full sum_i sum_j f_i * g_j matrix,
  i.e. write it as sum_i f_i * f_i + 2 sum_i sum_{j > i} f_i * f_j

Additionally, if we're doing a computation that at some point involves
both a square and a pairwise (as convolution-by-squaring can) we can try
to share work between them: if they happen to have the same FFT lengths,
then the splits and ffts will be the same.

Overall, for computing a square f * f the old code does:

- splitting of f (twice),
- #splits_f forward FFTs (twice)
- #splits_f * #splits_f backwards FFTs and filters

The new code does:

- splitting of f once
- #splits_f forward FFTs once
- #splits_f * (#splits_f + 1) / 2 backwards FFTs and filters

I.e. almost exactly half the amount of work (specifically, the addition
of the #splits_f / 2 in the last step)

For the case when a pairwise f * g and a square f * f happen at the same
time (such as f**7, at the second step of conv-by-squaring, we compute
both f**3 = f * (f**2) and f**4 = (f**2)**2 together) we save even more
work, since a lot of the work of the square can be stolen from the
pairwise operations. Steps:

- split f and g, once each
- #splits_f + #splits_g forward FFTs
- #splits_f * ((#splits_f + 1) / 2 + #splits_g) backwards FFTs and
  filters

(The original version does 2 additional splits of f, 2 #splits_f
additional forward FFTs and (#splits_f - 1) / 2 additional backwards
FFTs and filters. Using improved squaring reduces the overhead over the
merged version to 1 split of f and #splits_f forward FFTs.)